pyproject.toml
[project]
name = "orbital-risk-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "jupyterlab>=4.5.6",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "seaborn>=0.13.2",
    "spacetrack>=1.4.0",
]
